requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
groq
python-dotenv
